refactor structure of internal pacakge
