+++ packages/core/src/error/TRPCError.ts
@@ -1,5 +1,5 @@
-import { isObject } from '../internals/utils';
 import { TRPC_ERROR_CODE_KEY } from '../rpc/codes';
+import { isObject } from '../utilityFunctions';
 
 class UnknownCauseError extends Error {
   [key: string]: unknown;
