+++ packages/core/src/error/formatter.ts
@@ -1,9 +1,9 @@
+import { ProcedureType } from '../procedure';
 import {
   TRPC_ERROR_CODE_KEY,
   TRPC_ERROR_CODE_NUMBER,
   TRPCErrorShape,
 } from '../rpc';
-import { ProcedureType } from '../types';
 import { TRPCError } from './TRPCError';
 
 /**
