+++ packages/core/src/index.ts
@@ -10,10 +10,9 @@ export type {
   ProcedureRouterRecord,
   CreateRouterInner,
   Router,
-  RouterCaller,
   AnyRouterDef,
 } from './router';
-export { callProcedure, createCallerFactory } from './router';
+export { callProcedure, createRouterFactory } from './router';
 export type {
   Procedure,
   AnyProcedure,
@@ -30,22 +29,20 @@ export {
   experimental_standaloneMiddleware,
 } from './middleware';
 export type { MiddlewareFunction, MiddlewareBuilder } from './middleware';
+export * from './trpcConfig';
 export { initTRPC } from './initTRPC';
 export * from './types';
 export { type DefaultErrorShape } from './error/formatter';
-export type { RootConfig, AnyRootConfig } from './internals/config';
 export * from './transformer';
 
 export { mergeRouters } from './internals/mergeRouters';
 export * from './internals/procedureBuilder';
-export * from './internals/utils';
-export * from './internals/config';
+export * from './utilityFunctions';
 
 export * from './procedure';
 
 export * from './types';
 export * from './shared/createProxy';
-export * from './shared/jsonify';
 export * from './shared/transformTRPCResponse';
 
 // For `.d.ts` files https://github.com/trpc/trpc/issues/3943
@@ -53,4 +50,6 @@ export type { SerializeObject, Serialize } from './shared/serialize';
 
 export * from './shared/getErrorShape';
 
-export * from './shared/types';
+export * from './inference';
+export * from './utilityFunctions';
+export * from './utilityTypes';
